Move getTrainVehicleIdFromTrainEuropeanVehicleNumber to own file
